requirements.txt
boto3==1.23.10
mysqlclient==2.1.1
apache-airflow-providers-mysql==4.0.1
apache-airflow-providers-common-sql[pandas]